macros: share Flavor between ser and deser macros
