+++ scylla-macros/src/serialize/mod.rs
@@ -1,21 +1,2 @@
-use darling::FromMeta;
-
 pub(crate) mod row;
 pub(crate) mod value;
-
-#[derive(Copy, Clone, PartialEq, Eq, Default)]
-enum Flavor {
-    #[default]
-    MatchByName,
-    EnforceOrder,
-}
-
-impl FromMeta for Flavor {
-    fn from_string(value: &str) -> darling::Result<Self> {
-        match value {
-            "match_by_name" => Ok(Self::MatchByName),
-            "enforce_order" => Ok(Self::EnforceOrder),
-            _ => Err(darling::Error::unknown_value(value)),
-        }
-    }
-}
